Update README.md
A Spring Boot application to create a Timer with a unique id, creation time, start value which will update its value in every step time seconds.
Have implemented the below following functionalities:

Implement a REST based API server using Java 8 
and Spring Boot with following functions: 
• 1) A call to _create?start=<startval:integer>, 
step=<steptime:integer> should result in 
creation of a new go routine in the program 
that starts counting every steptime seconds. 
The response should return a unique identifier 
for that routine. 
• 2) A call to _check?id=<unique id:string> 
should return the current counter value, 
creation time and step time for that service. 
• 3) A call to _check should return same as (2) 
above but for all the unique identifiers present 
in the system. 
• 4) A call to _render should return an HTML 
page stating in tabular form, the unique 
identifiers and their current counts. 
• 5) A call to _clear/:id or _clear?id=<identifier: 
string> should clear the timer and cleanly exit 
setting the service status to stopped. 
• 6) A call to _pause?id=<identifier:string> 
should pause a given service if it exists, else 
return error. If it exists, and is a valid target, 
change the modifiedAt time to reflect this call’s 
response. 
• 7) Use dependency injection patterns to 
implement the above and discuss about the 
same. 
• 8) Document the above program in an 
appropriate README file. 
• 9) Discuss the use of interfaces in the above 
program if relevant. 
• 10) You can use any database backend if 
required by the above program for persistence
and state retrieval on abrupt kills. Discuss about 
persistency in the above context and update the 
README.